+++ backlog.md
@@ -2,10 +2,10 @@
 
 Current feature backlog of tasks to do
 
-- Support for more complex property names
-- Yaml Config Source support
 - dotenv config source support
 - Read config files from path
+- Yaml Config Source support
+- Support for more complex property names
 - Data Type converter support
 - more complex config builder
 - Expose Basic interface in python
